Adding product categories

+++ src/components/Layout/ProductPurchase.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { formatPrice } from "../../function.tsx/function";
+import { formatPrice } from "../../function/function";
 import { Autocomplete, TextField } from "@mui/material";
 import { api } from "../../service/api"
 import { dbProducts } from "../interface/dbInterfaces";
@@ -224,7 +224,7 @@ const ProductPurchase: React.FC = () => {
               type="number"
               className="text-center font-semibold text-sm md:text-base border rounded"
               value={item.quantity_purchased}
-              onChange={(e) => handleAmountChange(index, parseInt(e.target.value))}
+              onChange={(e) => handleAmountChange(index, parseFloat(e.target.value))}
               label="Quantity Pembelian"
             />
